-bouncy module, mark Bouncy Castle as an API dependency
BouncyPrivKey extends SecP256K1FieldElement and therefore has an API
dependency on Bouncy Castle. We should remove this dependency, but for
now we'll just mark this module as having an API dependency on Bouncy Castle.

Marking it as an API dependency allows us to remove declarations of direct
dependencies on Bouncy Castle from modules using secp256k1-bouncy. When
we remove the API dependency modules using secp256k1-bouncy will not
have to change their Gradle/Maven dependency declarations.  (But if they
actually depend on `SecP256K1FieldElement` that will have to change)

I plan on fixing this for 0.0.3 or maybe even sooner.