flake.lock: Update with `nix flake update`
This updates to Gradle 8.8, so I changed a comment in `flake.nix` that
referred to Gradle 8.7 to refer to Gradle 8.x.

The comment referencing JDK 22 on jextract is important, so let's leave
it for now. When Nix jextract is updated we will know and can update that
comment then.

Note that the Nix Gradle 8.x package will track the latest LTS JDK, so
_might_ update when JDK 25 is released, but of course we might be on
Gradle 9 by then.