flake.{nix,lock}: Update to JDK23 (temporarily uninclude jextract)
NixPkgs has removed JDK 22 as "unsupported" and now provides jdk23. The
jextract package has not been upgraded and is marked as `broken`.

Given that `jextract` is not needed to build `secp256k1-jdk` from source,
it is commented out for now.
(jextract is included in the `nix develop` environment for updating the generated
wrapper manually)

The flake.lock file is also updated to the latest hashes.