gradle.properties: javaToolchainVersion = 23
This sets the default toolchain to JDK 23. This does not change the supported
version of any of the outputs.

If for some reason you still need to build with JDK 22, you can use
the `-PjavaToolchainVersion=22` option to Gradle.

GitHub actions gradle.yml is also updated to no longer use the -P option
to override the toolchain version.