flake.nix: use JDK 23 to run devshell Gradle
Since our current config (flake.lock) is using Gradle 8.12.1,
we can (and should) use the `java` override property to install jdk23
rather than the `javaToolchains` property.

(This will change when we move to Gradle 8.14 and JDK 24, but for now
this is the simplest and fastest to install configuration)

We are also moving from temurin-bin-23 which was a workaround
for a problem that seems to have been fixed.

These changes reduce the number of JDKs that need to be installed
from 3 down to 1. (Nixpkgs `gradle` currently defaults to jdk21.)