Gradle: revert use of org.beryx.jar plugin
This reverts most of PR #140 / commit 2afefa09aac04af6b921fc8beeb07eeeffb535ce

The changes to Secp256k1Provider.java which removed use of Java 9+
features and APIs will remain, but the changes to the Gradle
files to use the org.beryx.jar plugin and the changes to configure it
during CI are reverted.

We still would like to be able to produce Java 8-compatible JARs for
both secp-api and secp-bouncy, but we will take a fresh look at the
best way to do so.