WIP: Graal native-image build of Java Schnorr Example
This is now rebased for JDK 25-ea and can compile because only native-image
is run on JDK 25 and it is not run directly from Gradle.

To build native image use:

./gradlew secp-examples-java:nativeCompile

Running does not work on aarch64-macOS and has not been tested on Linux.


Note: the Arena.ofConfined() change should probably not be merged to `master`,
but is currently necessary to work with native-image.

# Conflicts:
#	secp-examples-java/src/main/java/org/bitcoinj/secp/examples/ForeignRegistrationFeature.java
#	secp-integration-test/src/test/java/org/bitcoinj/secp/integration/package-info.java
#	secp256k1-examples-java/build.gradle