WIP: Graal native-image build of Java Schnorr Example
This is now rebased for JDK 25-ea and can compile because only native-image
is run on JDK 25 and it is not run directly from Gradle.

To build native image use:

./gradlew secp-examples-java:nativeCompileSchnorr

To run the native image see the added README section.

Note: This build currently does not work in the Nix devshell,
so you may need to disable it with `direnv deny`.