flake.nix: use nixpkgs-unstable, not nixos-unstable
nixpkgs-unstable is recommended for cross-platform builds, like ours.
Of course, we will migrate to nixpkgs-25.11 when it is released as its
gradle_9 should be Gradle 9.1+ and it should have OpenJDK 25 & GraalVM 25.

Today there was an update to nixos-unstable that did not have cached
binaries for many `darwin` packages and this broke some of our builds.
This is unlikely to happen if we are using nixpkgs-unstable.