Update build to javaToolchainVersion = 25
* gradle.properties: set javaToolchainVersion = 25
* settings.gradle: remove conditional for Java 25
* flake.nix: replace JDK 24 with graalvm 25, use nxpkgs-jdk-ea 25
* GHA gradle.yml: remove -PjavaToolchainVersion override
* .gitlab-ci.yml: remove -PjavaToolchainVersion override

Note that the changes in `flake.nix` switch from nixpkgs-unstable
to nixpkgs-jdk-ea/nixpkgs/jdk-ea-25, which is a fork of nixpkgs
that is nixpkgs/master + https://github.com/NixOS/nixpkgs/pull/434086

It is maintained by @msgilligan and nixpkgs/master is periodically
merged in. Once Gradle 9.1.0 is final, we can immediately switch
to nixpkgs/master and a few days later back to nixpkgs-unstable.