README.adoc: Restructure, Update for GraalVM 25, Homebrew, etc.
Now that we are requiring JDK 25, we can improve the documentation.

Reorganize the build instructions a little bit:

* Make them hierarchical
* Subsection for JDK installation
* Subsection for secp256k1 installation

We could add an AsciiDoctor `:toc:` but that can be considered
in a separate PR.

Document installing components with:

* Nix devShell
* APT
* Homebrew
* SDKMAN!

For Nix, the devShell is the preferred and documented way now (no 
`nix profile install` anymore)

Only Nix (barely!) has everything, but a combination of the others with
the Gradle Wrapper can work, too.

Document issues on macOS with System Integrity Protection.