Ecdsa.java: Fix ecPubKeySerialize flags for compressed serialization
Passing `2L` for flags is incorrect for compressed serialization. This
commit changes the value to 258 to match Ecdsa.kt which works correctly.

There is another pending PR that changes this to `pubkey.serialize(true)`,
which is simple, not error-prone, less verbose, and idiomatic.

And Issue #263 will remind us to properly document and define constants
for the `flags` parameter.